create AR, Avatar, Lighting Feature file
